requirements.txt
selenium
undetected-chromedriver
openai
pyyaml
flask
pytest
